Correction of native code to run with ffmpeg 2.2.3: Import new codecs Improve Packet (+correct free and alloc) Improve Container Improve StreamCoder with usage of swr to convert sample (any to/from FMT_16) Correct JNIHelper

+++ csrc/com/xuggle/xuggler/AudioSamples.h
@@ -72,7 +72,7 @@ namespace com { namespace xuggle { namespace xuggler
      *   That means that only getNumSamples()-startingSample
      *   samples are available in this AudioSamples collection.
      */
-    virtual short *getRawSamples(uint32_t startingSample);
+    virtual uint8_t *getRawSamples(uint32_t startingSample);
     
     /**
      * Called by decoder before decoding to ensure sufficient space
